split network tests into 3 files
